# swagger.json
{
  "swagger": "2.0",
  "info": { "title": "Football Proxy API", "version": "1.0" },
  "host": "https://football-proxy-xklx.onrender.com",
  "basePath": "/",
  "schemes": ["https"],
  "produces": ["application/json"],
  "paths": {
    "/competitions": {
      "get": {
        "operationId": "GetCompetition",
        "summary": "Get one competition by index",
        "parameters": [
          { "name": "index", "in": "query", "required": false, "type": "integer", "default": 0 }
        ],
        "responses": {
          "200": {
            "description": "One competition",
            "schema": {
              "type": "object",
              "properties": {
                "total":     { "type": "integer" },
                "index":     { "type": "integer" },
                "id":        { "type": "integer" },
                "name":      { "type": "string" },
                "code":      { "type": "string" },
                "type":      { "type": "string" },
                "emblem":    { "type": "string" },
                "area_name": { "type": "string" }
              }
            }
          }
        }
      }
    },
    "/standings": {
      "get": {
        "operationId": "GetStandingsRow",
        "summary": "Get one standings row by index",
        "parameters": [
          { "name": "code",     "in": "query", "required": true,  "type": "string" },
          { "name": "index",    "in": "query", "required": false, "type": "integer", "default": 0 },
          { "name": "season",   "in": "query", "required": false, "type": "integer" },
          { "name": "matchday", "in": "query", "required": false, "type": "integer" }
        ],
        "responses": {
          "200": {
            "description": "One standings row",
            "schema": {
              "type": "object",
              "properties": {
                "total":         { "type": "integer" },
                "index":         { "type": "integer" },
                "position":      { "type": "integer" },
                "team_id":       { "type": "integer" },
                "team_name":     { "type": "string" },
                "team_short":    { "type": "string" },
                "played":        { "type": "integer" },
                "won":           { "type": "integer" },
                "draw":          { "type": "integer" },
                "lost":          { "type": "integer" },
                "points":        { "type": "integer" },
                "goals_for":     { "type": "integer" },
                "goals_against": { "type": "integer" },
                "goal_diff":     { "type": "integer" }
              }
            }
          }
        }
      }
    },
    "/matches": {
      "get": {
        "operationId": "GetMatch",
        "summary": "Get one match by index",
        "parameters": [
          { "name": "code",     "in": "query", "required": true,  "type": "string" },
          { "name": "index",    "in": "query", "required": false, "type": "integer", "default": 0 },
          { "name": "season",   "in": "query", "required": false, "type": "integer" },
          { "name": "matchday", "in": "query", "required": false, "type": "integer" },
          { "name": "status",   "in": "query", "required": false, "type": "string" },
          { "name": "dateFrom", "in": "query", "required": false, "type": "string" },
          { "name": "dateTo",   "in": "query", "required": false, "type": "string" }
        ],
        "responses": {
          "200": {
            "description": "One match",
            "schema": {
              "type": "object",
              "properties": {
                "total":          { "type": "integer" },
                "index":          { "type": "integer" },
                "id":             { "type": "integer" },
                "utc_date":       { "type": "string" },
                "status":         { "type": "string" },
                "matchday":       { "type": "integer" },
                "home_team_id":   { "type": "integer" },
                "home_team_name": { "type": "string" },
                "away_team_id":   { "type": "integer" },
                "away_team_name": { "type": "string" },
                "winner":         { "type": "string" },
                "home_score":     { "type": "integer" },
                "away_score":     { "type": "integer" }
              }
            }
          }
        }
      }
    },
    "/scorers": {
      "get": {
        "operationId": "GetScorer",
        "summary": "Get one scorer by index",
        "parameters": [
          { "name": "code",   "in": "query", "required": true,  "type": "string" },
          { "name": "index",  "in": "query", "required": false, "type": "integer", "default": 0 },
          { "name": "season", "in": "query", "required": false, "type": "integer" },
          { "name": "limit",  "in": "query", "required": false, "type": "integer" }
        ],
        "responses": {
          "200": {
            "description": "One scorer",
            "schema": {
              "type": "object",
              "properties": {
                "total":               { "type": "integer" },
                "index":               { "type": "integer" },
                "player_id":           { "type": "integer" },
                "player_name":         { "type": "string" },
                "player_nationality":  { "type": "string" },
                "team_id":             { "type": "integer" },
                "team_name":           { "type": "string" },
                "goals":               { "type": "integer" },
                "assists":             { "type": "integer" },
                "penalties":           { "type": "integer" }
              }
            }
          }
        }
      }
    },
    "/teams": {
      "get": {
        "operationId": "GetTeam",
        "summary": "Get a team by ID",
        "parameters": [
          { "name": "id", "in": "query", "required": true, "type": "integer" }
        ],
        "responses": {
          "200": {
            "description": "Team details",
            "schema": {
              "type": "object",
              "properties": {
                "id":         { "type": "integer" },
                "name":       { "type": "string" },
                "short_name": { "type": "string" },
                "tla":        { "type": "string" },
                "crest":      { "type": "string" },
                "address":    { "type": "string" },
                "website":    { "type": "string" },
                "founded":    { "type": "integer" },
                "colors":     { "type": "string" },
                "venue":      { "type": "string" }
              }
            }
          }
        }
      }
    }
  }
}